2.0.0
Revamped:
- The way screensaver/lockscreens are checked on Linux
- Allow/block lists

- Cleaned up and unified the way the individual hooks are.

Now everything is caught and given a type. From there the cloak is checked, then the blocklist and the the allowlist.
Depending on whether this is an important/allowlisted person, is whether we'll get an e-mail or not.
Unless of course it's a PM or someone used our nick, then it's just checked if they're blocklisted.
Not really wanting to miss someone PMing/using our nick, considering it's probably important.

If you're not a fan of this approach, it's pretty simple to change.